+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/altcolor/AltColorResource.java
@@ -58,7 +58,7 @@ public class AltColorResource {
     try {
       descriptor.getAssetsToImport().add(AssetType.ALT_COLOR);
       descriptor.upload();
-      universalUploadService.importArchiveBasedAssets(descriptor, AssetType.ALT_COLOR);
+      universalUploadService.importArchiveBasedAssets(descriptor, null, AssetType.ALT_COLOR);
       gameService.resetUpdate(gameId, VpsDiffTypes.altColor);
       return descriptor;
     }